IE: change type of loop index from uchar to uint
Message from .L file: loop-statement induction variable must be 4-byte or 8-byte integer for array translation

+++ device/energy_ie.c
@@ -49,7 +49,7 @@ void energy_ie (
 	float interE = 0.0f;
 
 	// For each ligand atom
-	for (uchar atom1_id = 0; atom1_id < DockConst_num_of_atoms; atom1_id++)
+	for (uint atom1_id = 0; atom1_id < DockConst_num_of_atoms; atom1_id++)
 	{
 		char atom1_typeid = IA_IE_atom_types[atom1_id];
